easy wins: make BuildingID u32, not usize. maps from 680MB to 676MB. scenarios from 500MB to 458MB.
considered going down to u16, but pretty sure there are already maps
with >65k buildings

+++ game/src/common/warp.rs
@@ -106,7 +106,7 @@ fn inner_warp(ctx: &mut EventCtx, app: &mut App, line: &str) -> Option<Transitio
             }
             'l' => ID::Lane(LaneID(idx)),
             'i' => ID::Intersection(IntersectionID(idx)),
-            'b' => ID::Building(BuildingID(idx)),
+            'b' => ID::Building(BuildingID(idx as u32)),
             'a' => ID::Area(AreaID(idx)),
             'p' => ID::Pedestrian(PedestrianID(idx)),
             'P' => {
