+++ map_model/src/objects/building.rs
@@ -6,7 +6,7 @@ use std::fmt;
 
 // TODO reconsider pub usize. maybe outside world shouldnt know.
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct BuildingID(pub usize);
+pub struct BuildingID(pub u32);
 
 impl fmt::Display for BuildingID {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
